fix: revert to always rendering and swaping
The main loop should loop in 60hz for Shvav-8 to work properly (timers
and beepers).
I tried to render and swap buffers only when updated but
then the main loop won't wait for 60hz because it waits only when
swaping buffers.
I thought that maybe only swaping buffers without rendering would work
but it then flickers(one buffer is not updated).